zotify-main/requirements.txt
ffmpy
git+https://github.com/Googolplexed0/librespot-python.git@login5-auth
music_tag
Pillow
pkce
protobuf
pwinput
tabulate[widechars]
tqdm
python-telegram-bot>=20.0
